+++ react/FormInput-checkbox.tsx
@@ -4,5 +4,5 @@ import { CheckboxInput } from './components/CheckboxRenderer'
 import { BaseInputProps } from './typings/InputProps'
 
 export default function FormInput(props: BaseInputProps) {
-  return <CheckboxInput path={props.path} />
+  return <CheckboxInput pointer={props.pointer} />
 }