+++ react/__mocks__/vtex.render-runtime.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export const ExtensionPoint = props => {
+  return <p>ExtensionPoint: {props.id}</p>
+}