+++ react/components/InputRenderer.tsx
@@ -9,18 +9,18 @@ import {
 
 import { useFormattedError } from '../hooks/useErrorMessage'
 
-export const HiddenInput: FC<{ path: string }> = props => {
-  const inputObject = useHidden(props.path)
+export const HiddenInput: FC<{ pointer: string }> = props => {
+  const inputObject = useHidden(props.pointer)
   return <input {...inputObject.getInputProps()} />
 }
 
-export const PasswordInput: FC<{ path: string }> = props => {
-  const inputObject = usePassword(props.path)
+export const PasswordInput: FC<{ pointer: string }> = props => {
+  const inputObject = usePassword(props.pointer)
   return <InputRenderer inputObject={inputObject} />
 }
 
-export const RawInput: FC<{ path: string }> = props => {
-  const inputObject = useInput(props.path)
+export const RawInput: FC<{ pointer: string }> = props => {
+  const inputObject = useInput(props.pointer)
   return <InputRenderer inputObject={inputObject} />
 }