+++ react/hooks/useErrorMessage.tsx
@@ -22,8 +22,20 @@ const messages = defineMessages({
     id: 'store/form.error.minValue',
     defaultMessage: '',
   },
+  notFloat: {
+    id: 'store/form.error.notFloat',
+    defaultMessage: '',
+  },
+  notInteger: {
+    id: 'store/form.error.notInteger',
+    defaultMessage: '',
+  },
   pattern: {
     id: 'store/form.error.pattern',
+    defaultMessage: '',
+  },
+  notInEnum: {
+    id: 'store/form.error.notInEnum',
     defaultMessage: '',
   },
   multipleOf: {
@@ -46,18 +58,32 @@ export const useFormattedError = (message: ErrorMessage): string => {
     case ErrorTypes.required:
       return intl.formatMessage(messages.required)
     case ErrorTypes.maxLength:
-      return intl.formatMessage(messages.maxLength, { value: message.expected })
+      return intl.formatMessage(messages.maxLength, {
+        value: message.expected?.toString(),
+      })
     case ErrorTypes.minLength:
-      return intl.formatMessage(messages.minLength, { value: message.expected })
+      return intl.formatMessage(messages.minLength, {
+        value: message.expected?.toString(),
+      })
     case ErrorTypes.maxValue:
-      return intl.formatMessage(messages.maxValue, { value: message.expected })
+      return intl.formatMessage(messages.maxValue, {
+        value: message.expected?.toString(),
+      })
     case ErrorTypes.minValue:
-      return intl.formatMessage(messages.minValue, { value: message.expected })
+      return intl.formatMessage(messages.minValue, {
+        value: message.expected?.toString(),
+      })
+    case ErrorTypes.notFloat:
+      return intl.formatMessage(messages.notFloat)
+    case ErrorTypes.notInteger:
+      return intl.formatMessage(messages.notInteger)
+    case ErrorTypes.notInEnum:
+      return intl.formatMessage(messages.notInEnum)
     case ErrorTypes.pattern:
       return intl.formatMessage(messages.pattern)
     case ErrorTypes.multipleOf:
       return intl.formatMessage(messages.multipleOf, {
-        value: message.expected,
+        value: message.expected?.toString(),
       })
     default:
       return intl.formatMessage(messages.genericError)