+++ react/typings/FormProps.d.ts
@@ -1,0 +1,4 @@
+export type FormProps = {
+  entity: string
+  schema: string
+}